fix: rebase issues fixed in dashboard and router.
